Instruction: check operand type in debug mode

+++ tests/compiler/src/test_scenario_05.cpp
@@ -10,7 +10,7 @@ int main(int argc, char**) {
 
     std::array<X64Instruction, 2> instructions {{
         X64Instruction::make(0x0, Insn::PXOR_XMM_XMMM128, 1, XMM::XMM0, XMMM128{true, XMM::XMM0, {}}),
-        X64Instruction::make(0x4, Insn::JMP_U32, 1, 0x0),
+        X64Instruction::make(0x4, Insn::JMP_U32, 1, (u32)0x0),
     }};
 
     auto bb = cpu.createBasicBlock(instructions.data(), instructions.size());
